FIX : 좌표 더블로 수정
jira ref: #S12P21A204-210

+++ mobile/app/src/main/java/com/ssafy/jangan_mobile/model/PixelLatLng.kt
@@ -1,7 +1,7 @@
 package com.ssafy.jangan_mobile.model
 
 data class PixelLatLng(
-    val x: Int,
-    val y: Int,
+    val x: Double, // 경도 (Longitude)
+    val y: Double, // 위도 (Latitude)
     val floor: Int
 )
